# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalView": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "environmentId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkRequest"
          },
          "ruleId": {
            "type": "string"
          },
          "sandboxId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "denied",
              "dismissed"
            ],
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "environmentId",
          "kind",
          "subject",
          "attempts",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "Conn": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "description": "The destination as the sandbox asked for it",
            "type": "string"
          },
          "ended": {
            "format": "date-time",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "intercepted": {
            "description": "Studio handled the HTTP requests (proxy rules and hosts with secrets)",
            "type": "boolean"
          },
          "port": {
            "format": "int64",
            "type": "integer"
          },
          "received": {
            "description": "Bytes received by the sandbox",
            "format": "int64",
            "type": "integer"
          },
          "ruleId": {
            "type": "string"
          },
          "sent": {
            "description": "Bytes sent by the sandbox",
            "format": "int64",
            "type": "integer"
          },
          "source": {
            "description": "Where the host name came from",
            "enum": [
              "tls",
              "http",
              "dns",
              "socks",
              "ip"
            ],
            "type": "string"
          },
          "started": {
            "format": "date-time",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "pending",
              "open",
              "allowed",
              "denied",
              "undecided",
              "failed"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "started",
          "host",
          "port",
          "address",
          "source",
          "verdict",
          "sent",
          "received"
        ],
        "type": "object"
      },
...
      "DecideInBody": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "description": "allow or deny creates a rule; dismiss only closes the request",
            "enum": [
              "allow",
              "deny",
              "dismiss"
            ],
            "type": "string"
          },
          "host": {
            "description": "Host pattern the rule covers; empty means the requested host",
            "type": "string"
          },
          "ports": {
            "description": "Ports the rule covers; omitted means the default ports for the request, empty means any port",
            "items": {
              "format": "int64",
              "maximum": 65535,
              "minimum": 1,
              "type": "integer"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "scope": {
            "default": "environment",
            "description": "Whether the rule covers only the requesting sandbox",
            "enum": [
              "sandbox",
              "environment"
            ],
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "type": "object"
      },
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NetworkRequest": {
        "additionalProperties": false,
        "properties": {
          "host": {
            "type": "string"
          },
          "patterns": {
            "description": "Suggested host patterns, most specific first",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "port": {
            "format": "int64",
            "type": "integer"
          },
          "sandboxName": {
            "type": "string"
          }
        },
        "required": [
          "host",
          "port",
          "sandboxName",
          "patterns"
        ],
        "type": "object"
      },
...
      "Rule": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "allow",
              "proxy",
              "caddy",
              "deny"
            ],
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/RuleConfig"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "environmentId": {
            "type": "string"
          },
          "host": {
            "description": "example.com, *.example.com (includes example.com), or an IP address",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "ports": {
            "description": "Empty means any port",
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "sandboxId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "environmentId",
          "host",
          "ports",
          "action",
          "config",
          "note",
          "createdAt"
        ],
        "type": "object"
      },
...
      "RuleInput": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "description": "allow passes connections through untouched; proxy lets Studio handle the HTTP requests, to set headers; caddy hands them to a Caddyfile",
            "enum": [
              "allow",
              "proxy",
              "caddy",
              "deny"
            ],
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/RuleConfigInput"
          },
          "host": {
            "description": "example.com, *.example.com (includes example.com), or an IP address",
            "minLength": 1,
            "type": "string"
          },
          "note": {
            "maxLength": 500,
            "type": "string"
          },
          "ports": {
            "description": "Empty or omitted means any port",
            "items": {
              "format": "int64",
              "maximum": 65535,
              "minimum": 1,
              "type": "integer"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "sandboxId": {
            "description": "Limits the rule to one sandbox of this environment",
            "type": "string"
          }
        },
        "required": [
          "host",
          "action"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sandbox Studio",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/approvals": {
      "get": {
        "description": "Approvals of every environment, newest first. Pending ones feed the approval inbox.",
        "operationId": "listApprovals",
        "parameters": [
          {
            "description": "Statuses to list; all lists every status",
            "explode": false,
            "in": "query",
            "name": "status",
            "schema": {
              "default": "pending",
              "description": "Statuses to list; all lists every status",
              "enum": [
                "pending",
                "approved",
                "denied",
                "dismissed",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalView"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/environments/{env}/approvals": {
      "get": {
        "operationId": "listEnvironmentApprovals",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          },
          {
            "description": "Statuses to list; all lists every status",
            "explode": false,
            "in": "query",
            "name": "status",
            "schema": {
              "default": "all",
              "description": "Statuses to list; all lists every status",
              "enum": [
                "pending",
                "approved",
                "denied",
                "dismissed",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalView"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/environments/{env}/approvals/{id}/decide": {
      "post": {
        "operationId": "decideApproval",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          },
          {
            "description": "Approval ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Approval ID",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideInBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/environments/{env}/rules": {
      "get": {
        "operationId": "listRules",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "network"
        ]
      },
      "post": {
        "operationId": "createRule",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "network"
        ]
      }
    },
    "/api/environments/{env}/rules/{id}": {
      "delete": {
        "operationId": "deleteRule",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          },
          {
            "description": "Rule ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Rule ID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "network"
        ]
      },
      "put": {
        "operationId": "updateRule",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          },
          {
            "description": "Rule ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Rule ID",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "network"
        ]
      }
    },
...
    "/api/environments/{env}/sandboxes/{id}/connections": {
      "get": {
        "description": "The sandbox's most recent connections, newest first. The log is kept in memory only.",
        "operationId": "listConnections",
        "parameters": [
          {
            "description": "Environment ID",
            "in": "path",
            "name": "env",
            "required": true,
            "schema": {
              "description": "Environment ID",
              "type": "string"
            }
          },
          {
            "description": "Sandbox ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Sandbox ID",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Conn"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "network"
        ]
      }
    },
...
  }
}